Restructure file tree
