+++ pycompete/ensemble.py
